requirements.txt
numpy
jsonlines
matplotlib
scikit-learn
jupyterlab
torchaudio==0.6.0
librosa
pytorch-lightning
audiomentations
wandb
pyyaml
easydict
googledrivedownloader
glow